+++ docs/Changelog.md
@@ -2,12 +2,13 @@
 
 # 0.2.21 (next)
   - if `iterations` set, then duration for test will not be limited
-
+  - added `steps` option to execution settings
+  
 # 0.2.20
   - add `within` logic to timeframed pass-fail criterias
   - added `use-dns-cache-mgr` option.
   - default-domain option renamed to default-address (scheme, hostname, port are now parsed from this option).
-
+  
 # 0.2.19
   - fixed bug when in distributed tests VU count was not calculated properly.
   - auto-append `${__machineName()}` for thread names in distributed test